fix: Remove legacy Rarities contract
